+++ frontend/app/src/components/ConfirmDeleteModal.tsx
@@ -16,7 +16,9 @@ const ConfirmDeleteModal: React.FC<ConfirmDeleteModalProps> = ({
   isOpen ? (
     <View style={styles.overlay}>
       <View style={styles.modal}>
-        <Text style={styles.title}>Are you sure you want to delete entry?</Text>
+        <Text style={styles.title}>
+          Are you sure you want to delete the entry?
+        </Text>
         <Pressable style={styles.button} onPress={onConfirm}>
           <Text style={styles.buttonText}>Yes</Text>
         </Pressable>
